splite and use common class
